Move trackToCollision associator to common
