+++ Common/TableProducer/CMakeLists.txt
@@ -94,3 +94,8 @@ o2physics_add_dpl_workflow(fwdtrackextension
                                           O2::DetectorsBase
                                           O2::DetectorsCommonDataFormats
                     COMPONENT_NAME Analysis)
+
+o2physics_add_dpl_workflow(track-to-collision-associator
+                    SOURCES trackToCollisionAssociator.cxx
+                    PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore
+                    COMPONENT_NAME Analysis)
