+++ PWGHF/TableProducer/CMakeLists.txt
@@ -218,8 +218,3 @@ o2physics_add_dpl_workflow(candidate-selector-chic-to-jpsi-gamma
                     SOURCES candidateSelectorChicToJpsiGamma.cxx
                     PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore
                     COMPONENT_NAME Analysis)
-
-o2physics_add_dpl_workflow(track-to-collision-associator
-                    SOURCES trackToCollisionAssociator.cxx
-                    PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore
-                    COMPONENT_NAME Analysis)
